+++ src/core/internal/gt_moduleupgrader.cpp
@@ -39,8 +39,10 @@ public:
     }
 
     // logic
-    void runUpgradeRoutines(const GtVersionNumber& footprintVersion,
-                           const QStringList& files) const
+    void runUpgradeRoutines(const QString& projectName,
+                            bool saveWithLinkedFiles,
+                            const GtVersionNumber& footprintVersion,
+                            const QStringList& files) const
     {
         gtDebugId("module data upgrader") << "running upgrade routines...";
 
@@ -65,21 +67,18 @@ public:
                         continue;
                     }
 
-                    QDomDocument document;
-                    QString errorStr;
-                    int errorLine;
-                    int errorColumn;
+                    QStringList warnings;
+                    QDomDocument document = gt::xml::loadProjectXmlWithLinkedObjects(
+                        file.fileName(), &warnings);
 
-                    if (!gt::xml::readDomDocumentFromFile(file, document, true,
-                                                          &errorStr,
-                                                          &errorLine,
-                                                          &errorColumn))
+                    if (!warnings.isEmpty())
                     {
-                        gtErrorId("module data upgrader")
-                                  << "XML ERROR!" << " " << "line" << ": "
-                                  << errorLine << " " << "column" << ": "
-                                  << errorColumn << " -> " << errorStr;
+                        for (auto&& warn : warnings)
+                            gtErrorId("module data upgrader") << warn;
+                    }
 
+                    if (document.isNull() || document.documentElement().isNull())
+                    {
                         continue;
                     }
 
@@ -89,13 +88,22 @@ public:
                     upgradeRoutine.f(root, modData);
 
                     // save file
-                    // new ordered attribute stream writer algorithm
-                    if (!gt::xml::writeDomDocumentToFile(modData,
-                                                         document, true))
+                    const QFileInfo fi(modData);
+                    const QDir baseDir = fi.dir().absolutePath();
+
+                    auto saveType = saveWithLinkedFiles ?
+                                gt::xml::LinkFileSaveType::WithLinkedFiles :
+                                gt::xml::LinkFileSaveType::OneFile;
+
+                    QString error;
+                    if (!gt::xml::saveProjectXmlWithLinkedObjects(projectName,
+                                                                  document,
+                                                                  baseDir,
+                                                                  modData,
+                                                                  saveType,
+                                                                  &error))
                     {
-                        gtErrorId("module data upgrader")
-                                  << modData << QStringLiteral(": ")
-                                  << "Failed to save project data!";
+                        gtErrorId("module data upgrader") << error;
                     }
                 }
             }
@@ -176,7 +184,9 @@ GtModuleUpgrader::debugModuleConverter()
 }
 
 void
-GtModuleUpgrader::upgrade(const QMap<QString, GtVersionNumber>& moduleFootprint,
+GtModuleUpgrader::upgrade(const QString& projectName,
+                          bool saveWithLinkedFiles,
+                          const QMap<QString, GtVersionNumber>& moduleFootprint,
                           const QStringList& files) const
 {
     for (auto const& upgrader : m_pimpl->m_upgrader)
@@ -190,7 +200,8 @@ GtModuleUpgrader::upgrade(const QMap<QString, GtVersionNumber>& moduleFootprint,
 
             auto& upgradeHelper = upgrader.second;
 
-            upgradeHelper.runUpgradeRoutines(
+            upgradeHelper.runUpgradeRoutines(projectName,
+                        saveWithLinkedFiles,
                         moduleFootprint.value(QString::fromStdString(moduleId)),
                         files);
         }