+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Unknown tasks (corrseponding module did not load) are now displayed as Dummy Objects in the `GtTaskGroup` - #612
  - Added color property `GtColorProperty` to store color information in data models - #1329
  - New implementations for boundaries based on optional values to support property definitions - #1381
+ - `GtObjects` can now be stored as separate `*.gtobj.xml` files. Use `object.setSaveAsOwnFile(true);` to mark an object for linked-file creation. - #1419
 
 ### Changed
  - GTlab now requires C++17 for compilation and usage. - #1441